api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mangum==0.17.0
python-dotenv==1.0.0
httpx==0.25.2
PyGithub==2.1.1
openai==1.13.3
pydantic==2.5.2
